requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==5.0.7
django-decouple==2.1
django-dotenv==1.4.2
django-jazzmin==3.0.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
inflection==0.5.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.19.0
sqlparse==0.5.0
uritemplate==4.1.1
